Refactored TUI & Included Search Functionality (#70)
* Splitted TUI

Split the TUI into focused files so pkg/tui/app.go is now mostly the UI wiring/entrypoint while helpers live alongside it:

+ pkg/tui/state.go holds constants/enums and uiState/bucket object structs; formatters.go and table_helpers.go hold pure helpers.
+ View/controller logic now lives in services_view.go, buckets_view.go, and bucket_objects.go;
+ cross-cutting UI helpers went to dialogs.go, search.go, and auto_refresh.go.
+ app.go cleaned up accordingly, with redundant definitions removed and imports trimmed.

* Improved search functionality

+ Implemented support for search via "/" across Services, Buckets, Clusters and Details panels, including substrings.

* resolve conflict

* resolve conflict

---------
